pyproject.toml
[tool.poetry]
name = "worstcase"
version = "0.4.0"
description = "Worst case analysis and sensitivity studies. Extreme Value, Root-Sum-Square, Monte Carlo."
authors = ["amosborne <amosborne@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/amosborne/worstcase"

[tool.poetry.dependencies]
python = "^3.9"
networkx = "^2.5.1"
pyDOE = "^0.3.8"
Pint = "^0.20"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.12.1"
isort = "^5.8.0"
black = "^22"
flake9 = "^3.8.3"
pytest-randomly = "^3.8.0"
pytest-cov = "^2.12.0"
jupyterlab = "^3.0.15"
matplotlib = "^3.4.2"
jupyterlab-code-formatter = "^1.5.3"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
exclude = "*.ipynb_checkpoints*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
